rename utilities and cli scripts
