+++ faq-site/cli/lintData.js
@@ -1,7 +1,7 @@
 const fetch = require("node-fetch")
 const chalk = require("chalk")
 
-const utilities = require('./utilities')
+const utilities = require('../util/utilities')
 
 // run main code block
 validate()
