+++ faq-site/package.json
@@ -7,13 +7,15 @@
         "clean": "npx rimraf _site",
         "build": "npx @11ty/eleventy",
         "serve": "npx @11ty/eleventy --serve",
-        "archive-data": "node tools/archiveData.js",
-        "fetch-data": "npx cross-env AZURE_ENVIRONMENT=test node tools/updateData.js",
-        "lint-data": "node tools/lintData.js",
-        "list-changes": "node tools/listChanges.js",
-        "restore-local": "npx cross-env AZURE_ENVIRONMENT=test node tools/restoreTestKb.js",
-        "deploy-to-prod": "node tools/deployToProd.js",
-        "update-follow-ups": "node tools/fixFollowUpsOnLocalFaq.js"
+
+        "archive-data": "node cli/archiveData.js",
+        "fetch-data": "npx cross-env AZURE_ENVIRONMENT=test node cli/updateData.js",
+        "lint-data": "node cli/lintData.js",
+        "list-changes": "node cli/listChanges.js",
+
+        "restore-local": "npx cross-env AZURE_ENVIRONMENT=test node cli/restoreTestKb.js",
+        "deploy-to-prod": "node cli/deployToProd.js",
+        "update-follow-ups": "node cli/fixFollowUpsOnLocalFaq.js"
     },
     "keywords": [],
     "author": "",
