+++ faq-site/util/utilities.js
